% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api_response.R
\name{check_api_response}
\alias{check_api_response}
\title{Check api response}
\usage{
check_api_response(resp, content)
}
\arguments{
\item{resp}{A response}

\item{content}{The content}
}
\description{
A function that checks the api response, taken from Colin Gillespie's
rtypeform package
}
