% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_ingredients.R
\name{concat_ingredients}
\alias{concat_ingredients}
\title{Concatenating a multiple strings into one}
\usage{
concat_ingredients(no_of_ingredients, input)
}
\description{
Taking a set of ingredient inputs and concatenating them. They are collapsed
with a comma and no space
}
