% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter out search results that contain either wanted or unwanted
ingredients}
\usage{
filter_data(data, ingredients, wanted)
}
\arguments{
\item{data}{Data frame/tibble}

\item{ingredients}{A vector of ingredients}

\item{wanted}{Logical element dependent on whether the vectot of ingredients
are wanted or unwanted ingredients}
}
\description{
Filter out search results that contain either wanted or unwanted
ingredients
}
