% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api.R
\name{get_api}
\alias{get_api}
\title{Retrieve API key from Renviron}
\usage{
get_api(api = NULL)
}
\arguments{
\item{api}{Default \code{NULL}. Your private api key. If \code{api}
is \code{NULL}, the environment variable \code{Sys.getenv("getrecipes_api")}
is used.}
}
\description{
Taken from Colin Gillespie's rtypeform package
}
\details{
If the entry \code{"getrecipes_api"} exists in your
\code{.Renviron} return that value. Otherwise, raise an error.
}
