% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authorization.R
\name{get_authorization}
\alias{get_authorization}
\title{Get Authorziation}
\usage{
get_authorization(api)
}
\arguments{
\item{api}{An authentication key}
}
\description{
A function to get authorization taken from Colin Gillespie's rtypeform
package
}
