% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content.R
\name{get_content}
\alias{get_content}
\title{Parsing JSON text into a list containing recipe details}
\usage{
get_content(ingredients, type, pages)
}
\arguments{
\item{ingredients}{A string containing selected ingredients, separated by a
comma but with no space}

\item{type}{A character string containing the type of food/drink a user wishes
to make}

\item{pages}{The number of pages of search results}
}
\description{
Parsing JSON text into a list containing recipe details
}
