% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_response}
\alias{get_response}
\title{Getting a response from a server}
\usage{
get_response(api, url)
}
\arguments{
\item{api}{An application programming interface key to authenticate a user to an API}

\item{url}{The url of the webpage}
}
\description{
Taken from rtypeform package created by Colin Gillespie
}
