% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{get_url}
\alias{get_url}
\title{Extracting the URL of a recipe bunny webpage}
\usage{
get_url(ingredients, type, page, first_exist)
}
\arguments{
\item{ingredients}{A string containing each chosen ingredient, separated by a comma}

\item{type}{The desired type of food}

\item{page}{The chosen page of search results}

\item{first_exist}{Logical element determining if the first page exists - i.e.
if a server response is successful if p=1 is in the url}
}
\description{
We are taking the user parameter entries and parsing them into the URL link
}
