% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_pages2.R
\name{join_pages2}
\alias{join_pages2}
\title{Joining lists relating to different pages of search results}
\usage{
join_pages2(url, content, ingredients, type, pages, first_exist)
}
\description{
Joining lists relating to different pages of search results
}
