% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_join.R
\name{list_join}
\alias{list_join}
\title{List concatenation function to join recipe data}
\usage{
list_join(url, content)
}
\arguments{
\item{url}{An initialized vector}

\item{content}{A list containing recipe data}
}
\description{
List concatenation function to join recipe data
}
