% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_count.R
\name{page_count}
\alias{page_count}
\title{Function displaying text indicating the number of pages of search results}
\usage{
page_count(data, pages_wanted)
}
\arguments{
\item{data}{A data frame/tibble}

\item{pages_wanted}{Desired number of pages of search results}
}
\description{
Function displaying text indicating the number of pages of search results
}
