% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reached_limit.R
\name{reached_limit}
\alias{reached_limit}
\title{Exceeding search results indicator}
\usage{
reached_limit(first_exist, page)
}
\description{
Function that prints a message to indicate the user when they have exceeded
the number of search results
}
