% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_check.R
\name{recipe_check}
\alias{recipe_check}
\title{Check to see if a webpage contains recipes in the form of JSON text}
\usage{
recipe_check(url, content, i, first_exist)
}
\arguments{
\item{content}{A list containing fields of recipe information}

\item{i}{Page number selected}

\item{first_exist}{Logical element denoting whether or not the first page
of results contains recipes in the form of JSON text}

\item{api}{An authentication key}
}
\description{
This function checks to see whether a webpage contains recipes in the form of JSON data.
If so it will concatenate the list defined by the content argument with the new
list we have obtained from getting a response from the public API. If
}
