% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes_exist.R
\name{recipes_exist}
\alias{recipes_exist}
\title{A check for recipe data within a webpage}
\usage{
recipes_exist(url)
}
\arguments{
\item{url}{The url of a webpage}
}
\description{
A check for recipe data within a webpage
}
