% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_ui.R
\name{show_ui}
\alias{show_ui}
\title{Making a certain number of textboxes visible to the user}
\usage{
show_ui(no_of_ingredients, wanted)
}
\arguments{
\item{no_of_ingredients}{Number of ingredients}

\item{wanted}{Logical element dependent on whether the ingredients are wanted
or not}
}
\description{
Making a certain number of textboxes visible to the user
}
