% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_textboxes.R
\name{generate_textboxes}
\alias{generate_textboxes}
\title{Generating shiny textboxes}
\usage{
generate_textboxes(no_of_ingredients)
}
\arguments{
\item{no_of_ingredients}{Number of ingredients selected by the user}
}
\description{
The idea behind this function is to dynamically create a certain amount
of shiny UI textboxes based on the number of ingredients the user wants
to filter their search results by
}
