% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipe.R
\name{get_recipe}
\alias{get_recipe}
\title{Collating recipes based on user parameter entry}
\usage{
get_recipe(type, wanted_ingredients, pages, filter_all_ing = TRUE)
}
\description{
get_recipe() collects recipes that match the filters set by the user. It takes arguments
such as type of food/drink and the desired ingredients.
}
