% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_pages.R
\name{join_pages}
\alias{join_pages}
\title{Joining lists relating to different pages of search results}
\usage{
join_pages(url, content, ingredients, type, pages, first_exist)
}
\description{
Joining lists relating to different pages of search results
}
