% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_data.R
\name{manipulate_data}
\alias{manipulate_data}
\title{Further data manipulation on the recipes data}
\usage{
manipulate_data(recipe_data)
}
\arguments{
\item{recipe_data}{A data frame/tibble}
}
\description{
Further data manipulation on the recipes data
}
